+++ src/main/java/dev/abarmin/aml/dashboard/DashboardControllerAdvice.java
@@ -3,10 +3,12 @@ package dev.abarmin.aml.dashboard;
 import dev.abarmin.aml.config.AppConfiguration;
 import dev.abarmin.aml.dashboard.converter.BlockConverter;
 import dev.abarmin.aml.dashboard.converter.DashboardModelConverter;
+import dev.abarmin.aml.dashboard.converter.PageConverter;
 import dev.abarmin.aml.dashboard.domain.Block;
 import dev.abarmin.aml.dashboard.domain.Page;
 import dev.abarmin.aml.dashboard.model.BlockModel;
 import dev.abarmin.aml.dashboard.model.DashboardModel;
+import dev.abarmin.aml.dashboard.model.PageModel;
 import dev.abarmin.aml.dashboard.repository.BlockRepository;
 import dev.abarmin.aml.dashboard.repository.PageRepository;
 import dev.abarmin.aml.registration.domain.Profile;
@@ -17,10 +19,10 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 @ControllerAdvice(assignableTypes = {
   DashboardController.class,
+  PageUpdateHandler.class,
   BlockUpdateHandler.class
 })
 @RequiredArgsConstructor
@@ -31,6 +33,7 @@ public class DashboardControllerAdvice {
   private final BlockRepository blockRepository;
   private final BlockConverter blockConverter;
   private final AppConfiguration configuration;
+  private final PageConverter pageConverter;
 
   @ModelAttribute("model")
   public DashboardModel model(Authentication authentication,
@@ -41,13 +44,21 @@ public class DashboardControllerAdvice {
   }
 
   @ModelAttribute("currentBlock")
-  private BlockModel currentBlock(@PathVariable(value = "blockId", required = false) Long blockId) {
+  public BlockModel currentBlock(@PathVariable(value = "blockId", required = false) Long blockId) {
     if (Objects.isNull(blockId)) {
       return null;
     }
     final Block currentBlock = blockRepository.findById(blockId)
       .orElseThrow(() -> new IllegalArgumentException("Block not found"));
     return blockConverter.convert(currentBlock);
+  }
+
+  @ModelAttribute("currentPage")
+  public PageModel currentPage(Authentication authentication,
+                               @PathVariable(value = "pageId", required = false) Long pageId) {
+    final Profile profile = sessionService.getProfile(authentication);
+    final Page currentPage = getCurrentPage(profile, pageId);
+    return pageConverter.convert(currentPage);
   }
 
   @ModelAttribute("allowedImageTypes")