+++ src/main/java/dev/abarmin/aml/dashboard/block/BlockPropsSupport.java
@@ -1,7 +1,0 @@
-package dev.abarmin.aml.dashboard.block;
-
-import dev.abarmin.aml.dashboard.domain.BlockProps;
-
-public interface BlockPropsSupport<T extends BlockProps> {
-  T toProps();
-}