+++ src/main/java/dev/abarmin/aml/dashboard/domain/Page.java
@@ -17,4 +17,17 @@ public record Page(
   @Column("created_at")Instant createdAt,
   @Column("updated_at")Instant updatedAt
   ) {
+
+  public Page withProps(PageProps props) {
+    return new Page(
+      id,
+      profileId,
+      title,
+      isHome,
+      isDeleted,
+      props,
+      createdAt,
+      Instant.now()
+    );
+  }
 }