+++ src/main/java/dev/abarmin/aml/dashboard/model/PageModel.java
@@ -1,5 +1,7 @@
 package dev.abarmin.aml.dashboard.model;
 
+import dev.abarmin.aml.dashboard.domain.PageProps;
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -11,5 +13,8 @@ import java.util.Collection;
 public class PageModel {
   private long pageId;
   private String pageTitle;
+  @Valid
+  private PageProps pageProps;
+
   private Collection<BlockModel> pageBlocks = new ArrayList<>();
 }